requirements.txt
# Core
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.1

# Base de datos (si usas Postgres en el futuro)
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30

# Utilidades
python-dotenv==1.0.1

# WSGI server
gunicorn==22.0.0

# Pagos (Stripe)
stripe==5.8.0

# Seguridad / defensa
flask-limiter==3.7.0
PyJWT==2.10.1

# Bot de voz (Twilio)
twilio==9.2.3

# HTTP requests
requests==2.32.3
